requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0,<0.96.0
uvicorn[standard]>=0.22.0,<0.23.0
gunicorn>=20.1.0,<21.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.10.0,<1.11.0
psycopg2-binary>=2.9.5,<3.0.0  # PostgreSQL driver
pymysql>=1.0.2,<2.0.0  # MySQL driver (optional)

# Authentication
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.1.0  # For form data parsing

# Environment variables
python-dotenv>=1.0.0,<2.0.0

# Validation
pydantic>=1.10.7,<2.0.0
email-validator>=2.0.0,<3.0.0

# Scientific computing
numpy>=1.24.0,<2.0.0
scipy>=1.10.0,<2.0.0
sympy>=1.11.0,<2.0.0
matplotlib>=3.7.0,<4.0.0

# Caching
redis>=4.5.0,<5.0.0

# Performance optimizations
ujson>=5.7.0,<6.0.0  # Faster JSON parsing
orjson>=3.8.0,<4.0.0  # Even faster JSON parsing
uvloop>=0.17.0,<0.18.0  # Faster event loop
httptools>=0.5.0,<0.6.0  # Faster HTTP parsing

# Testing
pytest>=7.3.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
httpx>=0.24.0,<0.25.0  # For testing HTTP clients

# Logging
loguru>=0.7.0,<0.8.0  # Better logging

# Firebase (optional)
# firebase-admin>=6.1.0,<7.0.0 